converted big integers to unsigned integers for relationships

+++ database/migrations/2021_11_08_145204_create_measurements_table.php
@@ -14,7 +14,7 @@ class CreateMeasurementsTable extends Migration
             $table->string('type');
             $table->unsignedDouble('ratio');
             $table->unsignedDouble('rounding_precision');
-            $table->bigInteger('measurement_category_id')->nullable();
+            $table->unsignedInteger('measurement_category_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
