+++ database/migrations/2022_02_04_073444_create_product_categories_table.php
@@ -11,7 +11,7 @@ class CreateProductCategoriesTable extends Migration
         Schema::create('product_categories', function (Blueprint $table) {
             $table->id();
             $table->string('category');
-            $table->bigInteger('parent_product_category_id')->nullable();
+            $table->unsignedInteger('parent_product_category_id')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
