+++ database/migrations/2022_02_17_235908_create_locations_table.php
@@ -11,7 +11,7 @@ class CreateLocationsTable extends Migration
         Schema::create('locations', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigInteger('parent_location_id')->nullable();
+            $table->unsignedInteger('parent_location_id')->nullable();
             $table->boolean('is_a_scrap_location')->default(false);
             $table->boolean('is_a_return_location')->default(false);
             $table->string('type');
