+++ database/migrations/2022_02_28_035442_create_transfers_table.php
@@ -11,17 +11,17 @@ class CreateTransfersTable extends Migration
         Schema::create('transfers', function (Blueprint $table) {
             $table->id();
             $table->string('reference')->nullable();
-            $table->bigInteger('contact_id')->nullable();
-            $table->bigInteger('operation_type_id')->nullable();
-            $table->bigInteger('source_location_id')->nullable();
-            $table->bigInteger('destination_location_id')->nullable();
+            $table->unsignedInteger('contact_id')->nullable();
+            $table->unsignedInteger('operation_type_id')->nullable();
+            $table->unsignedInteger('source_location_id')->nullable();
+            $table->unsignedInteger('destination_location_id')->nullable();
             $table->dateTime('scheduled_date');
             $table->string('source_document')->nullable();
             $table->string('tracking_reference')->nullable();
             $table->unsignedDouble('weight')->nullable();
             $table->unsignedDouble('weight_for_shipping')->nullable();
             $table->string('shipping_policy')->nullable();
-            $table->bigInteger('responsible_id')->nullable();
+            $table->unsignedInteger('responsible_id')->nullable();
             $table->string('note')->nullable();
             $table->string('status')->nullable();
             $table->string('shipping_method')->nullable();
