+++ database/migrations/2022_11_08_145254_create_access_rights_table.php
@@ -11,7 +11,7 @@ class CreateAccessRightsTable extends Migration
         Schema::create('access_rights', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->bigInteger('group_id')->nullable();
+            $table->unsignedInteger('group_id')->nullable();
             $table->boolean('read_access')->default(false);
             $table->boolean('write_access')->default(false);
             $table->boolean('create_access')->default(false);
